+++ src/lang-xquery/test/uk/co/reecedunn/intellij/plugin/intellij/tests/lang/XQueryParameterInfoHandlerTest.kt
@@ -308,5 +308,47 @@ private class XQueryParameterInfoHandlerTest : ParserTestCase() {
             assertThat(ui.highlightStart, `is`(-1))
             assertThat(ui.highlightEnd, `is`(-1))
         }
+
+        @Test
+        @DisplayName("parameters; first parameter highlighted")
+        fun parameters_first() {
+            val context = createParameterInfoContext("replace(1, 2, 3)", 8)
+            val function = XPathParameterInfoHandler.findElementForParameterInfo(context)
+
+            val ui = MockParameterInfoUIContext<XPathFunctionCall>(function)
+            ui.currentParameterIndex = 0
+
+            XPathParameterInfoHandler.updateUI(context.itemsToShow?.first() as XPathFunctionDeclaration, ui)
+            assertThat(ui.currentParameterIndex, `is`(0))
+            assertThat(ui.parameterOwner, `is`(sameInstance(function)))
+            assertThat(ui.isSingleOverload, `is`(false))
+            assertThat(ui.isSingleParameterInfo, `is`(false))
+            assertThat(ui.isUIComponentEnabled, `is`(false))
+
+            assertThat(ui.text, `is`("\$input as xs:string?, \$pattern as xs:string, \$replacement as xs:string"))
+            assertThat(ui.highlightStart, `is`(0))
+            assertThat(ui.highlightEnd, `is`(20))
+        }
+
+        @Test
+        @DisplayName("parameters; last parameter highlighted")
+        fun parameters_last() {
+            val context = createParameterInfoContext("replace(1, 2, 3)", 14)
+            val function = XPathParameterInfoHandler.findElementForParameterInfo(context)
+
+            val ui = MockParameterInfoUIContext<XPathFunctionCall>(function)
+            ui.currentParameterIndex = 2
+
+            XPathParameterInfoHandler.updateUI(context.itemsToShow?.first() as XPathFunctionDeclaration, ui)
+            assertThat(ui.currentParameterIndex, `is`(2))
+            assertThat(ui.parameterOwner, `is`(sameInstance(function)))
+            assertThat(ui.isSingleOverload, `is`(false))
+            assertThat(ui.isSingleParameterInfo, `is`(false))
+            assertThat(ui.isUIComponentEnabled, `is`(false))
+
+            assertThat(ui.text, `is`("\$input as xs:string?, \$pattern as xs:string, \$replacement as xs:string"))
+            assertThat(ui.highlightStart, `is`(45))
+            assertThat(ui.highlightEnd, `is`(70))
+        }
     }
 }